Speed up git cherry-pick operations on large changes
Similar to change Ib652dd9a58f8fb18354eb00e889212b72cafe402, this
replaces the gitpython diff calls with cgit commands.  This should
be faster when performing diffs on large changes or large numbers
of files.  The git diff command is minimal, so it should be fast
for cgit to calculate and will produce minimal output.

Additionally, we may be able to avoid one of the two diff operations
depending on the outcome of the first, so we also make that
optimization.

Change-Id: I67ddee7b15d83e1d0756b7874f907be26760170a