Clean up tox remnants
We don't test Zuul with tox these days, and modern tox versions
don't work with it. We had one lingering tox job defined because
zuul-client testing relied on it once upon a time, but that is no
longer the case either.

Update TESTING.rst to correctly refer to nox now in examples, so as
not to mislead anyone.

Change-Id: I4825a2e0cedc69f971d9c2b6961e928a11dddb4d