Record missing info in job-output.json
We use an Ansible callback plugin to record information about playbooks,
plays, and tasks executed by zuul jobs. This only works if Ansible exits
cleanly and has a chance to record playbook stats. If Ansible does not
exit cleanly (for example due to a timeout) then the entire playbook
goes missing in the json file.

Address this by having the executor detect such situations and add basic
info for a missing playbook to the json file. At this point it really
just points out that something wrong and the playbook may have timed
out, but we could potentially expand this to record more data in the
future.

Doing this should make the information available in the zuul web UI
console tab. Hopefully making it clearer to users why something isn't
quite as they expect in the logs.

Change-Id: I2bf446f2a05d3107d880b2b5132475fb4ead923e