Use global S3 endpoint to get bucket location
In some cases - e.g. when the S3 bucket is owned by a different AWS
account - we can't use the regional scoped boto3 client to determine the
location as it will fail with the following (misleading) error message:

    botocore.exceptions.ClientError: An error occurred (AccessDenied)
    when calling the GetBucketLocation operation: User: ... is not
    authorized to perform: s3:GetBucketLocation on resource: ... because
    no resource-based policy allows the s3:GetBucketLocation action

Instead, we need to use the global S3 endpoint, which also works cross
account given the correct bucket and IAM permissions.

Change-Id: Ic0ab56925c73f9795ee9d22b2f7f45d9aa74acee