Update gerrit change_url format
Modern Gerrits use a format for the change url that includes the
project name and change number.  Since the web interface expects
users to use this form, the UX around visiting changes which are
not visible to logged-in users is better.  While the short version
we previously used works if the user has access to the change,
if they do not, then Gerrit may return a nearly-empty page.  The
longer form of the URL produces a better error message and a
button to sign in.

We should therefore prefer the newer version of the URL that Gerrit
itself typically generates.  This form was introduced in 2.15, so
we only enable it for versions greater than that.

Change-Id: I260cc90519dbb81a0945b46adeed292b08eb13a9