Add static driver
This is part of nodepool-in-zuul.

This adds a static driver.

Static nodes are registered and de-registered on configuration
updates.  Each unique host-port-user combination is registered as a
Zuul Node in ZK with a fixed node UUID based on the uuid5 algorithm.
This algorithm produces stable outputs from these inputs.  They will
not conflict with the random uuid4 UUIDs that dynamic nodes use.  By
using a stable ID, we simplify conflict resolution with multiple
launchers (multiple launchers might try to register the same nodes at
the same time, but the unique IDs will ensure we have exactly the
right set).

The registration methods are in the launcher rather than the driver
because they re-use some internal launcher methods for convenience and
consistency.

Static nodes are always reused, and may optionally host slots.

A new node state, DEREGISTER, is used to mark nodes as needing to be
deregistered.  It is similar to OUTDATED.

If a static node encounters an ssh connection error, it will still be
re-used, but only after a 5 minute delay to avoid busy loops.  This is
implemented in the delete state machine for the static driver, which
will either set the state to BUILDING if the node should go back into
service after an ssh failure, or do nothing if the node is to be
deleted.

The routine to assign an unassigned node was previously only run when
we initially accepted a nodeset request.  It is now also run after the
launcher encounters an error for a node in the request and attempts to
replace the failed node.  This was necessary for this change since
static nodes are never created, therefore the only way to replace a
failed static node in a request is to find another ready one.  This
change improves performance for dynamic drivers as well since it
increases the chances of using a ready node.

This differs slightly from the spec in that nodes are only able to
be added to providers, not sections.  That's largely because as
implemented, the behavior can be well-defined and consistent with the
behavior we currently have for dynamic drivers.  If we put node
definitions on sections, we would need to once again allow nodes
to float between providers and we just stopped doing that for
the dynamic drivers.  It should be possible to make that change if
necessary in the future.

Some incidental changes included:

* Since the static driver immediately registers nodes on startup,
  the nodescan test fakes are updated to provide a fixture we can
  enable in the test setup method.

* Some assertions in the test_cloud_driver tests are corrected
  (not directly related to this change).

* The image.type configuration option is pulled up out of the aws
  and openstack schemas and placed in the base configuration.  They
  are always required for images, but make no sense to put in the
  static driver's source code.

* The "slot" attribute is removed from the Instance class since that
  doesn't make sense (an instance corresponds to a main node, not a
  particular subnode slot).  It should not have been copied from the
  instance to the node after state machine creation (but this was
  unlikely to cause an error).

* The tests did not exercise the bestEffortLockAndSetNodeState method
  in the code path where the node was already locked; the static node
  tests do exercise it and a bug with the nullcontext manager is
  corrected.

* A previous change corrected an issue with deserializing nods and
  assignments; some tests are updated to check for the request_id to
  ensure correct behavior.

Change-Id: I3d9c45fa91da2d5651e575f513b667054c9a29a6