Fix providernode assignment upgrade
A recent change to move assignment information to a subnode did
not handle ZK schema upgrades gracefully (which, as a matter of
policy, is okay because this only affects NIZ which is currently
under heavy development).  But the particular upgrade error is
persistent for any existing nodes.  Therefore we need an actual
upgrade path so that OpenDev can clean up its data without manual
intervention.  (As an alternative, we could manually intervene,
but I think this patch is a good foundation for future changes.)

The issue is this: we moved three attributes from the ProviderNode
to a child zk node called "assignment".  Any existing ProviderNodes
with these attributes will still have those attributes serialized,
and when we try to deserialize them, we will encounter an error
when trying to update the object.  The correct behavior is to
dynamically create the assignment node when deserializing the
ProviderNode, and then remove the attributes from the ProviderNode's
data.

Typically the ProviderNode is deserialized in the cache playback
(and all our affected nodes with old data will be deserialized on
startup).  In this location, we do not provide a ZKContext object,
but we need one in order to create a new ZKObject for the assignment
node.  To address this, we update ZuulTreeCache and its users to
supply a ZKContext object to the deserialization method so that
objects may perform other ZK operations if necessary.

This does mean that on startup, there will now be a slight inefficiency
in that we will read in the assignment znode twice: once when the
ProviderNode is loaded, then again when the cache loads its assignment
node.  Because this only happens on startup, we'll just accept the
additional read ops for now.

There is one location where we use a ZKObject but still don't pass
a ZKContext to the deserialize method: the tenant state object (used
for pausing tenant processing) does not have easy access to a ZKContext.
Since this is only one simple object and we know it won't use the
context, we continue to omit it from that call site.  We can add one
in the future if we need to.

This change also corrects a potentially serious error in the sequencing
of deserializing ProviderNode objects.  We need to set the UUID as
soon as possible when deserializing so that subnodes (such as the
assignment node) can reference the correct UUID when generating their
paths.

Change-Id: I1b001af52a9b799c48ac65d0d8352db789957bc6