Further fix to node assignment upgrade
The fix for the upgrade of nodes to using assignment child znodes
was incomplete.  In many (most?) cases, when we deserialize the
node, we will not write it back out to ZK, which means that the
upgrade process may repeat on the next load.  If we already
created an assignment znode, we should ignore the resulting error
from trying to create it again.

That does mean that we may continue to assign nodes to "ghost"
requests, but the system should unassign them after it detects that
situation.  And this is expected only to affect a few nodes in
opendev during this transition.

Change-Id: I91442f1fb9d616c4a2067a690796e348c059f40b