Add tenant ssh keys
With Nodepool, we had the option of setting "allowed-labels" for
tenants in order to restrict certain nodepool labels to certain
Zuul tenants.  With NIZ, we adopt a more Zuul-centric approach where
access control to node resources is determined by which resources
are present in Zuul's configuration for each tenant.

We have a number of options for addressing this with dynamic (cloud)
providers by controlling whether tenants can create their own sections
and providers.  But in the case of the static driver, if a tenant
is allowed to create a provider at all, then it can simply duplicate
the static node list from another tenant and gain access to those
nodes.

To give operators a tool to address this, we introduce the tenant
ssh key, which is halfway between the global "nodepool" ssh key and
the per-project ssh key.  Each tenant in the system gets one and
it is automatically generated.  An operator can use this key instead
of the global ssh key on any static nodes that should only be
available to certain tenants.

Unlike the per-project key, this key is available for all pipelines,
not only post-review pipelines.

Change-Id: I733937d3b697463484ab47d97564a54abf9e0ca1