Don't redact ssh key info in freeze job
The current output of the job request data structure which is sent
to the exceutor does not include the actual ssh keys to be used for
the build, but rather only a list of dictionaries with information
pointing to those keys.  We do not need to redact that information
when we supply it to the user via the freeze-job API.  Remove the
redaction so that users may inspect the actual list of keys that
will be supplied to the job.

The dictionaries include information such as: tenant name (for
tenant keys), and connection/project names (for project keys).

Change-Id: I5d1f06f8162b6c4c70263e5d78df4c1f3baa6bd7