Fix test_launcher_image_expire_failed_upload
At one point in this test it waited until there are 2 artifacts of
an image, however, the launcher can start deleting one of them
while that check is happening (since it happens right after we
start work to replace one of them).  Instead, wait for one artifact,
which will cause us to wait until that artifact is deleted -- we
know this because it's after the completion of the build that will
have created it, so we know we will at least have the second artifact
in the registry, so we start waiting either when there are 2 artifacts
about to reduce to 1 artifact, or after we have already reduced to
1 artifact.

Since we now exit the loop with only the second artifact, we need to
save the uuid of the first artifact earlier in the test.

Change-Id: Id6177392e48b46db698b01de98353049ca61c38a