Add image export/import commands
This is part of Nodepool in Zuul.

This adds a pair of zuul admin client commands that will export
and import the image data from ZooKeeper.  Similar commands exist
in nodepool today, so that admins can recover from a loss of ZK
data if they have image builds that are costly to recreate.  This
is similar to the key export/import commands.

Change-Id: Ic328ded3a9d141eea89cd056af3830ab2f795188