Update no_log in win_command and friends
Thi win_command and win_shell commands did not honor the no_log
setting when processing the live log streaming.  It is honored
by Ansible after the modules return control back to Ansible,
so therefore it has been filtered out of the JSON log output
(ie, job-output.json), but output would appear in the live
log stream and in the job-output.txt file.

The Windows modules do not have access to the task object like
the Posix modules do, therefore we need to pass an extra argument
to the Windows modules to indicate whether the no_log setting
is set on the task.  This argument will now also be possed to
the Posix modules, but they will ignore it.  We could switch
to using it in the future if we want; but that's not important --
the values will always be the same.

To keep things simple (and also to anticipate a future improvement
in handling large amounts of output on a no_log task), we still
run all the stream handling, processing, and threading code needed
to stream logs, but we simply skip the step where we write anything
to disk.  It is expected that we will make the same change to the
Posix modules soon.

Change-Id: Iddba67937762680796c71e0daf308e5c0a9b0c09