Fix image upload validation race conditions
So far we immediately sent the image-validate event as soon as an upload
completed (as part of the endpoint upload job). However, we only updated
and unlocked the upload after all futures had completed.

Depending on the image import/upload methods used, this could stall the
pipeline processor until the upload lock was released, since the
reporter blocks on acquiring the upload lock. This happens whenever a
single image validation is faster to complete than all the uploads for a
particular image build artifact.

The problem is fixed by updating and unlocking image uploads as they
complete. The image-validate event is now send just before we are about
to release the lock.

A second issue could trigger a data race between the launcher and a
scheduler reporting a successful validation. The problem here is a
scheduler refreshing the image upload without holding a lock, which
could lead to an exception (BadVersionError) when the scheduler tried
updating the validated flag.

This is addressed by refreshing the image upload after we've acquired
the lock.

Change-Id: I35e93d2257bd65b7010d14293c2aff17e45302a4