Add assertions about no_log and registered results
When we set no_log for a task, we prohibit stdout and stderr from
being included in any registered results variable for the task.

Since this is a behavior change from Ansible, set some assertions
in tests so that we know if the behavior is consistent over time.

(Note, it would probably be okay for the output to be registered,
but the mechanism we use to suppress it in the JSON output log
means that it is not available.  We could consider changing that
in the future.)

Change-Id: I00f5c165d9817ae2cec787ba1b653e463a745dc8