Add label aliases for static nodes
The spec, as well as the current nodepool implementation, allows
static nodes to have more than one label associated with them in
order to support all kinds of combinations of hardware.  For example,
a user may have 2 static nodes, either of which can run cpu-based
tests, but only one of which has a gpu attached.

The current NIZ implementation assumes a node has a single label
associated with it, and derives some of its configuration and
behavior from that single label.  Therefore, to support this use
case, we continue with the 1:1 association of a node to a label,
but we also add the ability for users to indicate that other labels
may be used to refer to the node.  These other labels are "aliases"
and are not used to derive behavioral aspects of the node, but
instead may only be used to request the node.

They must still be top-level label objects and attached to the
provider.  This may not be ideal (since, even more so than the
standard label for the static node, absolutely nothing about
the image and flavor configuration of the label aliases are
necessary, but they are still required).  We may relax that
requirement in a later change.

Change-Id: Ibd3d854413675b524aade8048a3c36b03fe6cc74