Annotate image upload logs with iba/upload uuid
In order to correlate image upload related logs more easily we will
annotate them with the image build artifact or upload UUID.

To allow this, the helper function for getting an annotated logger was
modified to accept the new annotations.

Since we are now adding more arguments to that function, we make all but
the event keyword-only arguments. At the same time we allow event to be
None, so we don't need to pass it in case we are in a context that
doesn't have an event.

Existing usage of `get_annotated_logger()` without an event were updated
and in one case (in the executor context) the event argument was added.

Change-Id: I5c9cb2b04c335314dc922f1ce55b7a491747b168