Replace slash when normalizing metrics in launcher
In addition to '.' and ':' we should also replace the '/' character when
normalizing statsd names. The reason is that this might cause problems
with some statsd implementations that store the metrics as filesystem
paths.

Unfortunately we can not do this globally right now, as we have existing
metrics that don't sanitize the slash.

Change-Id: I224fe95b50569b1f5c7e02b158cc67715dc4e6cf