Fix job-output.json fallback content
It turns out that my testing method for hacking custom job json output
into the web ui was too simple and a layer too deep. At the layer above
we also read and processed the job-output.json file and expected the
play stats to be present there. I've since modified my testing routine
to look things up one level higher in `receiveBuildOutput` and been able
to test this more robustly.

With better testing it is clear that the stats dict is required for web
ui processing. However, since we're in fallback mode we don't actually
know how many tasks succeeded and failed so we return an empty dict.
Additionally, we can provide information on whether or not the playbook
is trusted so we do that too. The index is converted to a string to
better match what Ansible does.

Finally, if we have timeout information we add a play duration to our
fallback content that attempts to capture the runtime for the entire
playbook. Since we don't actually know how long each playbook was
running this is a best effrot. We also don't bother capturing additional
timestamps and instead rely on the watchdog if it is present to
calculate the duration. This means that some playbooks may not include
duration info. If we feel that adding additional timestamps to track
durations for every playbook is worthwhile this change can be updated to
do that.

Change-Id: I7f698e800d5962f62421c204af0a67b4fd82a1d0