Fix missing update in node reassignment
When assigning unassigned nodes to a request with a failed node,
an indentation error prevented the launcher from updating its
in-memory list of nodes, and it continued operating as if the
failed node was assigned to the request; this would cause it to
fail the request immediately after assigning a ready node.

Correct the indentation error and add a test for this behavior.

Change-Id: Ic6ac55db5246a25cb6fe2e65853a9f5e7b720312