Fix BadVersionError in static tests
Several tests in the static driver lock a node and change its state
after it becomes READY, but it may be possible for the launcher
to make another change to the node between the transition to READY
and when it is unlocked.  Make sure we refresh the node after
locking it so we have the latest version.

Change-Id: I33300d33d9a1177f68408b3d110d46d2c0cbc456