Fix race in quickstart test
Update the "Wait for Zuul" check to use the components api to
verify that the scheduler is running (which means its config is
fully loaded) rather than the status endpoint.  This can avoid
a race in the secondary tutorials where the status endpoint could
return a cached status page before the scheduler has started and
won the ZK election for event subscription.

Change-Id: I2208d16d75448a47a3ff57c51dfb6cb49cd2e934