Fix image creation race
We delete image build artifacts when there are no uploads.  To
avoid a race condition when initially creating the IBA and uploads,
we need to do the following:

* Get the list of IBAs that are marked ready or deleting
* Get all the uploads
* Consider deleting the IBAs from the first list iff they have
  no uploads.

This sequence should work because the scheduler does the following:

* Create an IBA with state=None
* Create all the uploads for that IBA
* Update the IBA to state=ready

This ensures that if we get the list of ready IBAs before the list
of uploads, we will only consider IBAs where all the uploads exist.

This change corrects that order.

There is one possible exception: because the IBAs and uploads use
two different ZKObjectCache objects, it could be the case that the
upload cache is backlogged; to avoid that, wait for the upload cache
to progress at least past the point at which we fixed our list of
active IBAs before getting the list of uploads.

Also, update one debug log entry to be more clear.

Change-Id: I9e7006c4f9a06d101558825657c91b147be1b633