Force provider subnode class registration
Since subnodes are not a separate driver, they don't get imported as
part of the connection registry initialization.

This means that web and executors are not aware of the
`SubnodeProviderNode` subclass as it is not implicitly imported. This
can e.g. lead to the following exception:

ERROR zuul.AnsibleJob: [e: ...] [build: ...] Error locking nodeset <NodeSet ...>
Traceback (most recent call last):
  File ""/opt/zuul/lib/python3.13/site-packages/zuul/zk/zkobject.py"", line 714, in _fromRaw
    klass = cls._subclasses[subclass_id]
            ~~~~~~~~~~~~~~~^^^^^^^^^^^^^
KeyError: b'subnode'
During handling of the above exception, another exception occurred:
Traceback (most recent call last):
  File ""/opt/zuul/lib/python3.13/site-packages/zuul/executor/server.py"", line 1363, in lockNodes
    self.executor_server.launcher.acceptNodeset(
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^
        self.nodeset_request, self.nodeset)
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ""/opt/zuul/lib/python3.13/site-packages/zuul/launcher/client.py"", line 146, in acceptNodeset
    provider_node = ProviderNode.fromZK(
        ctx, path=ProviderNode._getPath(node_id))
  File ""/opt/zuul/lib/python3.13/site-packages/zuul/zk/zkobject.py"", line 698, in fromZK
    return cls._fromRaw(context, raw_data, zstat, None, **kw)
           ~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ""/opt/zuul/lib/python3.13/site-packages/zuul/zk/zkobject.py"", line 716, in _fromRaw
    raise RuntimeError(f""Unknown subclass id: {subclass_id}"")
RuntimeError: Unknown subclass id: b'subnode'"

To fix this we'll import the subnode class as part of the connection
registry.

Change-Id: I51e01cf20647a8168edb6849373d368c9f4f11b4