Launcher: honor provider when selecting ready nodes
When we are able to use unassigned (ready or building) nodes, we
should still only use those nodes from the provider that the
selectProviders method chose for us.  That is because it takes
into account node locality for the request, so we don't use
ready nodes from different providers.

This change also fixes a bug where we accumulated but did not
emit debug statements for a request if we were able to fulfill
the request using only unassigned nodes.

Change-Id: Ieef9a2b1b77abab1165f201979c38f1e0af691a3