Add k8s driver
This implements the k8s driver based on nodepool with some changes.

It is designed to accomodate all of the variations we currently
support in nodepool with a single code-base:

* k8s pods in namespaces
* k8s bare namespaces
* openshift pods in projects
* openshift bare namespaces

It does not currently implement openshift pods in a fixed namespace
(ie, the "openshiftpods" driver), because it is unclear whether we
should continue to support that in Zuul.  If we decide to do so,
it can be easily extended for that.

It improves the data structure that we pass to the executor, so that
we don't override the "connection_port" attribute.

This structure includes the list of containers in a pod.  In the future,
we can update the executor to add containers individually to the
inventory so they are individually addressable.

Change-Id: I0148f33d5ff7bc51e3e0549ac6f86dc87cc53ee8