Allow disabling ZKObject loading error logs
Besides the pipeline summary we started using nested ZKObjects for cases
where we could not modify a locked object, e.g. when revising a nodeset
request or claiming a building provider node.

However, when trying to load those objects we will always log a ZKObject
loading error.

With this change we add a class-level flag that allows us to not log
those errors in case the node doesn't exist.

The flag is applied to the following ZKObjects:
* PipelineChangeList
* PipelineSummary
* NodesetRequestRevision
* ProviderNodeAssignment

Change-Id: Ie533e7a7619da16eefe52c170eea4be0275fc69d