Wake up launcher main event loop every 60s
In Zuul setups without a lot of activity (be it in general or just
during certain times of the day), we can end up with expired nodes that
are not cleaned up.

The reason is that we only wake up the launcher if there is any
node/request activity in Zookeeper.

To fix this, we will wake up a launcher at least every 60s to trigger
any pending cleanup actions.

Change-Id: I3651794cd0c439ba4ea7e89b99e11b9dcdb3aef5