Fix OIDC secret config errors in zuul-web
When loading OIDC secrets we use the globals (system attributes) from
the scheduler reference of the parse-context.

However, the scheduler is set to None in zuul-web. Since we already pass
the system attributes there separately we can use this reference.

This change also removes the scheduler reference from the parse context
and instead passes it the globals directly.

Change-Id: Ic67cc48a1d5786340ef3103cae5b69ccc9206b14