Delete unlocked used nodes
Launchers should delete unlocked used nodes; that's the typical
end of the node lifecycle.  However, the actual end-of-life check
that we used for nodes was if its request was missing and we had
not previously decided to do something else with the node (like
make it available for reuse).  This was almost as good as checking
that it was used and unlocked, except that another recent bug
caused a node to be unassigned right as the request was being
fulfilled, which meant it got used without a request assignment.

That bug will be addressed separately, but this change corrects
the cleanup method so that no matter how a node gets into an
unlocked+used state, we will clean it up.

Change-Id: Ib8d9e1f00378cf11ebce09fa1d2bc29d83ca0fe1