Ensure requests referred to in assignments are up to date
The following sequence can happen with two launchers:

A: notice a new request
A: assign ready node to the request
B: notice node is assigned to a request that is not in its cache
B: unassign node from missing request
A: complete request
B: notice the request

This is because the node and request caches are maintained separately.
While all of the events arrive in the correct order from ZK, they
are processed by different threads in the launcher, and it is possible
for them to be out of sync.

To address this, if we ever see a node refer to a request that is not
in our cache, we will wait for the cache to completely synchronize with
the events it has received before proceeding.  Because any event
associating a node with a request should arrive after the event notifying
us that the request was created, that means that by the time we start
processing an assigned node, the request cache should *at least* have
the request creation event in its queue.  We simply need to wait for
any pending events to propagate all the way through the playback queue,
then we can be sure the request really does not exist.

The waitForSync method is updated to wait for a sentinel through both
the event and playback queue, and also to migrate sentinel events to
the new queues if a disconnection happens during the wait.

Change-Id: I31184f9d8499f541a24603da0ec4c9c3d268d984