Add support for zk watch zxids
ZooKeeper 3.9 introduced support for including the zxid of the
transaction that triggered a watch event.  Add support to our
vendored kazoo to retrieve that information.  This will help
us reduce race conditions when we check that our caches are
up to date.

Upstream PR: https://github.com/python-zk/kazoo/pull/765

Change-Id: I13654deade71e9e936e6cbc855b849c2e00a64fe