Ensure requests referred to in assignments are up to date
The following sequence can happen with two launchers:

A: notice a new request
A: assign ready node to the request
B: notice node is assigned to a request that is not in its cache
B: unassign node from missing request
A: complete request
B: notice the request

This is because the node and request caches are maintained separately.
While all of the events arrive in the correct order from ZK, they
are processed by different threads in the launcher, and it is possible
for them to be out of sync.

To address this, if we ever see a node refer to a request that is not
in our cache, we will wait for the cache to at least synchronize up
to the transaction id when the request was created before proceeding.

Change-Id: I31184f9d8499f541a24603da0ec4c9c3d268d984