Revert "Make Gerrit Connection Replication Aware"
Gerrit does not provide a universal way of detecting when all
replicas have received any git updates for an event as well as
completed any index updates required for that event.

It does provide replication completed events for push replication,
which is good enough for that case assuming that the index delay is
not noticeable.  But if indexing does take longer, it may not
be sufficient.  Additionally, processing the replication-done events
through ZK can add considerable load, and even more if a problem
causes Gerrit to re-attempt those replication events.

Any other method of replication provides no notification at all,
including (unexpectedly), pull replication, which has corresponding
events, but they are not emitted over the event streaming protocol,
and instead are only used internally by Gerrit.

Given all the above, it does not seem feasible to automate this in
Zuul and we should just expose a tunable configuration option.

This change removes the push replication handling, a subsequent
change will add a new tunable configuration option.

This reverts commit c0746040e30e9a65e742bc728d5bdde5ce8103e6.
This reverts commit f6a83a832f7a630d1518807eae2ce0b6d990193e.

Change-Id: Ifd591bf4ce25c189367e617b873ce647dd10c289