Add replication_delay to Gerrit driver
This makes the Gerrit replication delay (previously called propogation
delay internally in the code) configurable by users.  Since it is
typically not needed, we can set the default value to 0, but users
of replicated systems can increase it as necessary to match their
environment.

The main issue is that Gerrit does not provide a universal way of
detecting when all replicas have received any git updates for an
event as well as completed any index updates required for that event.

It does provide replication completed events for push replication,
which is good enough for that case assuming that the index delay is
not noticeable.  But if indexing does take longer, it may not
be sufficient.  Additionally, processing the replication-done events
through ZK can add considerable load, and even more if a problem
causes Gerrit to re-attempt those replication events.

Any other method of replication provides no notification at all,
including (unexpectedly), pull replication, which has corresponding
events, but they are not emitted over the event streaming protocol,
and instead are only used internally by Gerrit.

Given all the above, it does not seem feasible to automate this in
Zuul and we should just expose a tunable configuration option.

Change-Id: I89991a82a7aa869d24021d9b1413858d4157eca6