Launcher; use waitForSync as a fallback for older ZK
We want to use ZooKeeper 3.9 with the launcher due to the new
feature where zxids are included in watch events which makes our
cache synchronization easier and faster.  But since this is a
bit of a surprise to those of us testing the launcher, let's
include a less-efficient fallback in case an older version of
ZK is used, so that we can all work to upgrade ZK in parallel
to continues NIZ development.

Change-Id: I82af91729ada6529dd98a4a7f6c13cca5a794e6d