updateCommitDependencies: add some debug logs to explain why a depends-on might not be used
Add some helpful messages in debug to understand why a depends-on
in the commit header might not end up being used.

For example, gitlab's web UI will properly parse a reference to
another Merge Request in the form '<project>!<id>' and display the
proper hyperlink. Zuul on the other hand expects a proper URL.
The new debug messages will make it clear when such a case is
encountered.

Change-Id: I3a347d862e08c70ea92a0c2dd22dd75f3b7369ad