Fix gerrit event delay for single event
On a nearly idle system, if we receive a single event from gerrit,
we will delay an additional 10 seconds beyond the configured
replication delay.  This is because the following sequence occurs
(directed by BaseThreadPoolEventConnector._dispatchEventsMain;
let's assume a configured replication_delay of 10 seconds):

* We receive the event and add it to the ZK event queue
* _disptachEventsMain calls GerritEventConnector.dispatchEvents
* dispatchEvents calculates the delay, sleeps for the delay, and
  returns the delay value without processing the event
* _dispatchEventsMaincalls self._dispatcher_wake_event.wait(delay or 10)

That last call causes a further 10 seconds to elapse before we process
the event.

To correct this, we will remove the timeout from the event wait in the
last item above, and we will also set the event if we previously delayed.
This means that we should never perform the wait if we also performed
a sleep.

This means that we will delay processing of previously created forward
queue futures, but it increases the amount of time we spend in a simple
sleep call, which may help us avoid running through the dispatch loop
frequently with no work to do while we wait for the delay to elapse.

There were some leftover references to the GerritEventProcessor.delay
which were unused and are removed.

Change-Id: Ifd72c58c725147fac105840fce3e2cc873584383