Fix gerrit event delay for single event
On a nearly idle system, if we receive a single event from gerrit,
we will delay an additional 10 seconds beyond the configured
replication delay.  This is because the following sequence occurs
(directed by BaseThreadPoolEventConnector._dispatchEventsMain;
let's assume a configured replication_delay of 10 seconds):

* We receive the event and add it to the ZK event queue
* _disptachEventsMain calls GerritEventConnector.dispatchEvents
* dispatchEvents calculates the delay, sleeps for the delay, and
  returns the delay value without processing the event
* _dispatchEventsMaincalls self._dispatcher_wake_event.wait(delay or 10)

That last call causes a further 10 seconds to elapse before we process
the event.

To correct this, we will remove the sleep call in the Gerrit driver.
This allows us to return the calculated delay and continue to process
any accumulated event futures while we wait for it to elapse.

We also change the wait(delay or 10) call to wait(delay) because if
no delay is configured, we don't need a timeout.  We know whether
there are any events to process (via either the wake event or the
delay value), or whether there are any events in the forward queue
(from the line above).

There were some leftover references to the GerritEventProcessor.delay
which were unused and are removed.

Change-Id: I751cc73a999f920ebf3e9b448b44ddfffa59d86a