Add system events viewer to web ui
This allows users to click on the reconfiguration label on the
status page (the one that says "reconfigured X ago") in order
to navigate to a new page at /t/tenant/system-events which will
show the reconfiguration events for that tenant.

This page is constructed like the config-errors page so that we
can add filtering to it in the future if that proves useful.

Change-Id: I2400613fbb21369b7061c0681347379c87ddd1ae